add inference code for speehless
